+++ includes/admin/views/header.php
@@ -5,12 +5,6 @@
 	<?php endif; ?>
 
 	<aside class="alg-sidebar">
-		<div class="alg-sidebar__brand">
-			<div class="alg-sidebar__brandwrapper">
-				<img src="<?php echo ALGOLIA_WOOCOMMERCE_URL; ?>assets/img/algolia-logo.png">
-				<span>For WooCommerce</span>
-			</div>
-		</div>
 		<ul class="alg-sidebar__list">
 			<li class="alg-sidebar__link <?php if ( ! isset( $_GET['tab'] ) || $_GET['tab'] === 'pages'): ?>alg-sidebar__link--active<?php endif; ?>">
 				<a href="admin.php?page=algolia-woocommerce">
@@ -44,6 +38,8 @@
 				</a>
 			</li>
 		</ul>
+
+    <button class="algolia-reindex-button button button-primary" style="margin: 18px; width: 84%;" data-index="posts_product"><?php _e('Re-index products', 'algolia-woocommerce'); ?></button>
 	</aside>
 
 	<main class="alg-main">