+++ includes/indexing.php
@@ -153,3 +153,15 @@ function aw_product_index_settings( array $settings ) {
 
 add_filter( 'algolia_posts_product_index_settings', 'aw_product_index_settings' );
 
+add_filter( 'algolia_post_images_sizes', function( $sizes ) {
+    $sizes[] = 'shop_catalog';
+
+    return array_unique( $sizes );
+});
+
+
+add_filter( 'algolia_get_synced_indices_ids', function( $ids ) {
+    $ids[] = 'posts_product';
+
+    return array_unique( $ids );
+} );